% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_auth.R
\name{cloud_auth}
\alias{cloud_auth}
\title{cloud_auth}
\usage{
cloud_auth(reset_keys = FALSE)
}
\arguments{
\item{reset_keys}{Override existing keys.}
}
\value{
Nothing. Keys are stored in keyring.
}
\description{
\code{cloud_auth()} gets the user's credentials and stores them securely in
keyring.
}
