% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-auth.R
\name{cloud_auth_pwd}
\alias{cloud_auth_pwd}
\title{cloud_auth_pwd}
\usage{
cloud_auth_pwd()
}
\value{
user password as string.
}
\description{
Used to return a stored cloudstor password, or request the user set one.
Not a user-facing function.
}
