% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_auth.R
\name{cloud_auth_user}
\alias{cloud_auth_user}
\title{cloud_auth_user}
\usage{
cloud_auth_user()
}
\value{
user id as string.
}
\description{
Used to return a stored Cloudstor username, or request the user set one.
Not a user-facing function.
}
\keyword{internal}
