% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_get}
\alias{cloud_get}
\title{cloud_get}
\usage{
cloud_get(
  user = cloud_auth_user(),
  password = cloud_auth_pwd(),
  cloud_address,
  dest
)
}
\arguments{
\item{user}{cloudstor user name}

\item{password}{cloudstor password}

\item{cloud_address}{cloudstor webDav address to file of interest}

\item{dest}{destination for saving the file}
}
\value{

}
\description{
cloud_get
}
