% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_get.R
\name{cloud_get_path}
\alias{cloud_get_path}
\title{Get file from a path}
\usage{
cloud_get_path(
  cloud_address,
  path,
  dest,
  user,
  password,
  open_file,
  reset = FALSE,
  save = TRUE,
  ...
)
}
\arguments{
\item{cloud_address}{Address to use for fetching.}

\item{path}{The path to the file on cloudstor.}

\item{dest}{The optional path to save the file.}

\item{user}{Cloudstor username}

\item{password}{Cloudstor password}

\item{open_file}{If TRUE, open the file using rio.
Else, returns the file path.}

\item{reset}{Passed to \code{\link[=get_handle]{get_handle()}}}

\item{save}{Passed to \code{\link[=get_handle]{get_handle()}}}

\item{\dots}{pass additional arguments to \code{rio::import()}}
}
\value{
The file object or folder path is returned, depending on \code{open_file}
}
\description{
The engine function for \code{\link[=cloud_get]{cloud_get()}}, when the provided path is a private
file path. Not intended to be invoked directly.
}
\seealso{
\code{\link[=cloud_get_url]{cloud_get_url()}}
}
\keyword{internal}
