% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_get.R
\name{cloud_get_url}
\alias{cloud_get_url}
\title{Get file(s) from a public URL}
\usage{
cloud_get_url(cloud_address, url, dest, user, password, open_file, ...)
}
\arguments{
\item{cloud_address}{Address to use for fetching.}

\item{url}{The url to the file or folder.}

\item{dest}{The optional path to save the file(s).}

\item{user}{The 'key' for the public link. This is the last part of the
public URL}

\item{password}{The password for the public link. If there is no password,
use \code{NULL} or \code{""}.}

\item{open_file}{If TRUE, open the file using rio.
Else, returns the file path. This is ignored if the URL is to a folder.}

\item{\dots}{pass additional arguments to \code{rio::import()}}
}
\value{
One of: file object, file path, or a folder path.
}
\description{
The engine function for \code{\link[=cloud_get]{cloud_get()}}, when the provided path is a public
URL. Not intended to be invoked directly.
}
\seealso{
\code{\link[=cloud_get_path]{cloud_get_path()}}
}
