% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_meta}
\alias{cloud_meta}
\title{cloud_meta}
\usage{
cloud_meta(
  user = cloud_auth_user(),
  password = cloud_auth_pwd(),
  cloud_address = "https://cloudstor.aarnet.edu.au/plus/remote.php/webdav/"
)
}
\arguments{
\item{user}{Your cloudstor username}

\item{password}{Your cloudstor password}

\item{cloud_address}{cloudstor webDav address to the location you want to save on cloudstor}
}
\value{

}
\description{
cloud_meta
}
