% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_put}
\alias{cloud_put}
\title{Title}
\usage{
cloud_put(
  file_name,
  local_file,
  user,
  password,
  cloud_address = "https://cloudstor.aarnet.edu.au/plus/remote.php/webdav/"
)
}
\arguments{
\item{file_name}{What you want to call the file on cloudstor}

\item{local_file}{Where the file is located}

\item{user}{Your cloudstor username}

\item{password}{Your cloudstor password}

\item{cloud_address}{cloudstor webDav address to the location you want to save on cloudstor}
}
\value{

}
\description{
Title
}
