% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstoR-package.R
\docType{package}
\name{cloudstoR-package}
\alias{cloudstoR}
\alias{cloudstoR-package}
\title{cloudstoR: Simplifies access to cloudstor api.}
\description{
This pakage aims to provides access for reading and writing to cloudstor via their webDav API.
}
\author{
\strong{Maintainer}: Philip Parker \email{philip.parker@acu.edu.au} (\href{https://orcid.org/0000-0002-4604-8566}{ORCID})

Authors:
\itemize{
  \item Timothy Churches
  \item Taren Sanders \email{taren.sanders@acu.edu.au} (\href{https://orcid.org/0000-0002-4504-6008}{ORCID})
}

}
\keyword{internal}
