% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_or_folder}
\alias{file_or_folder}
\title{Check if metadata indicates a file or folder}
\usage{
file_or_folder(metadata)
}
\arguments{
\item{metadata}{metadata as returned by \code{\link[=cloud_meta]{cloud_meta()}}.}
}
\value{
One of c("folder", "file")
}
\description{
Used internally to check if a path should be treated as a folder or a single
file.
}
\keyword{internal}
