% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{get_cloud_address}
\alias{get_cloud_address}
\title{get_cloud_address}
\usage{
get_cloud_address(path)
}
\arguments{
\item{path}{path to file or folder}
}
\value{
encoded url as string
}
\description{
Return the full cloud_address path. Not a user facing function.
}
