% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_handle}
\alias{get_handle}
\title{get_handle}
\usage{
get_handle(user, password, reset = FALSE, save = TRUE)
}
\arguments{
\item{user}{Cloudstor username or file key (for public links)}

\item{password}{Cloudstor password or file password (for public links)}

\item{reset}{remove the existing authentication and handle}

\item{save}{save the new handle for reuse}
}
\value{
curl handle object
}
\description{
Return a handle for CURL to use. Not a user facing function
}
\keyword{internal}
