% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_dest_path}
\alias{make_dest_path}
\title{Generate the path for saving files}
\usage{
make_dest_path(dest, path = NULL)
}
\arguments{
\item{dest}{The optional destination provided by the user.}

\item{path}{The path to the file on cloudstor.}
}
\value{
A file path that can be used for saving files.
}
\description{
Used internally to determine where files should be saved.
}
\keyword{internal}
