% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_or_url}
\alias{path_or_url}
\title{Check if using a file path or a URL}
\usage{
path_or_url(item)
}
\arguments{
\item{item}{The user-provided string.}
}
\value{
"path" or "url"
}
\description{
Checks if the string provided to \code{path} in the \verb{cloud_*} functions are a
Cloudstor path or a URL (indicating a public file or folder).
Not a user facing function.
}
\keyword{internal}
