% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-auth.R
\name{cloud_auth}
\alias{cloud_auth}
\title{cloud_auth}
\usage{
cloud_auth(reset_keys = FALSE)
}
\arguments{
\item{reset_keys}{override existing keys}
}
\value{
Nothing. Keys are stored in Keyring.
}
\description{
Get users credentials and store them securely in keyring.
}
