% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_browse}
\alias{cloud_browse}
\title{cloud_browse}
\usage{
cloud_browse(path = "", user = cloud_auth_user(), password = cloud_auth_pwd())
}
\arguments{
\item{path}{initial path to start the search}

\item{user}{Your cloudstor username}

\item{password}{Your cloudstor password}
}
\value{
the last file path
}
\description{
Navigate the folder tree interactively.
}
