% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_get}
\alias{cloud_get}
\title{cloud_get}
\usage{
cloud_get(
  path,
  dest = NULL,
  user = cloud_auth_user(),
  password = cloud_auth_pwd(),
  open_file = TRUE,
  ...
)
}
\arguments{
\item{path}{The path to file or folder.}

\item{dest}{The destination for saving the file.}

\item{user}{Cloudstor user name}

\item{password}{Cloudstor password}

\item{open_file}{If TRUE, open the file using rio. Else, returns the file path}

\item{\dots}{pass additional arguments to \code{rio::import()}}
}
\value{
The file object or folder path is returned, depending on \code{open_file}
}
\description{
\code{cloud_list()} downloads a file from a Cloudstor folder. The file is opened
and read into R using rio, or optionally the file path is returned.
}
