% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_get.R
\name{cloud_get}
\alias{cloud_get}
\title{Fetch a file or files from Cloudstor}
\usage{
cloud_get(
  path,
  dest = NULL,
  user = NULL,
  password = NULL,
  open_file = TRUE,
  ...
)
}
\arguments{
\item{path}{The path to file or URL to file or folder.}

\item{dest}{The destination for saving the file.}

\item{user}{Cloudstor username. Not needed for a URL.}

\item{password}{Cloudstor password or optional password to public link.}

\item{open_file}{If TRUE, open the file using rio.
Else, returns the file path. This is ignored if the URL is a folder.}

\item{\dots}{pass additional arguments to \code{rio::import()}}
}
\value{
The file object or folder path is returned, depending on \code{open_file}
}
\description{
Download a file or files from a Cloudstor folder. The file is opened and
read into R using rio, or optionally the file path is returned. You can
provide either a path to a private Cloudstor file, or a public URL (with
optional password) to either a single file or a folder.
}
