% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_get}
\alias{cloud_get}
\title{cloud_get}
\usage{
cloud_get(
  path = "",
  user = cloud_auth_user(),
  password = cloud_auth_pwd(),
  dest
)
}
\arguments{
\item{path}{path to file or folder}

\item{user}{cloudstor user name}

\item{password}{cloudstor password}

\item{dest}{destination for saving the file}
}
\value{

}
\description{
Download a file from a cloudstor folder and open it in R.
}
