% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_get}
\alias{cloud_get}
\title{cloud_get}
\usage{
cloud_get(
  path,
  dest = NULL,
  user = cloud_auth_user(),
  password = cloud_auth_pwd(),
  open_file = TRUE
)
}
\arguments{
\item{path}{path to file or folder}

\item{dest}{destination for saving the file}

\item{user}{cloudstor user name}

\item{password}{cloudstor password}

\item{open_file}{if TRUE, open the file using rio. Else, returns the file path}
}
\value{
the file object or folder path, depending on open_file
}
\description{
Download a file from a cloudstor folder. The file is opened
and read into R using rio, or optionally the file path is returned.
}
