% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_list.R
\name{cloud_list}
\alias{cloud_list}
\title{cloud_list}
\usage{
cloud_list(path = "", user = cloud_auth_user(), password = cloud_auth_pwd())
}
\arguments{
\item{path}{The path to file or folder.}

\item{user}{Cloudstor user name.}

\item{password}{Cloudstor password.}
}
\value{
A list of files and folders.
}
\description{
Return a list of the files located in a folder.
}
