% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_list}
\alias{cloud_list}
\title{cloud_list}
\usage{
cloud_list(
  user = cloud_auth_user(),
  password = cloud_auth_pwd(),
  cloud_address = "https://cloudstor.aarnet.edu.au/plus/remote.php/webdav/"
)
}
\arguments{
\item{user}{cloudstor user name}

\item{password}{cloudstor password}

\item{cloud_address}{the cloudstor webdav address to your file/folder of interest}
}
\value{

}
\description{
cloud_list
}
