% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_list}
\alias{cloud_list}
\title{cloud_list}
\usage{
cloud_list(path = "", user = cloud_auth_user(), password = cloud_auth_pwd())
}
\arguments{
\item{path}{path to file or folder}

\item{user}{cloudstor user name}

\item{password}{cloudstor password}
}
\value{
a list of files and folders
}
\description{
cloud_list
}
