% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_meta.R
\name{cloud_meta}
\alias{cloud_meta}
\title{cloud_meta}
\usage{
cloud_meta(path = "", user = NULL, password = NULL, fetch_type = NULL)
}
\arguments{
\item{path}{The path to file or folder.}

\item{user}{Your Cloudstor username}

\item{password}{Your Cloudstor password}

\item{fetch_type}{One of \code{c("path", "url")}, for if \code{path} is a private path
or public URL.}
}
\value{
A data.frame of the file and folder metadata is returned.
}
\description{
Return the metadata for a file or folder. This can be useful
for checking if a file has been modified.
}
