% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_meta}
\alias{cloud_meta}
\title{cloud_meta}
\usage{
cloud_meta(path = "", user = cloud_auth_user(), password = cloud_auth_pwd())
}
\arguments{
\item{path}{The path to file or folder.}

\item{user}{Your Cloudstor username}

\item{password}{Your Cloudstor password}
}
\value{
A data.frame of the file and folder metadata is returned.
}
\description{
\code{cloud_meta()} returns the metadata for a file or folder. This can be useful
for checking if a file has been modified.
}
