% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_meta}
\alias{cloud_meta}
\title{cloud_meta}
\usage{
cloud_meta(path = "", user = cloud_auth_user(), password = cloud_auth_pwd())
}
\arguments{
\item{path}{path to file or folder}

\item{user}{Your cloudstor username}

\item{password}{Your cloudstor password}
}
\value{
a data.frame of the file and folder metadata
}
\description{
Return a data.frame of the metadata for a file or folder
}
