% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_put.R
\name{cloud_put}
\alias{cloud_put}
\title{cloud_put}
\usage{
cloud_put(
  local_file,
  path = "",
  file_name = basename(local_file),
  user = cloud_auth_user(),
  password = cloud_auth_pwd()
)
}
\arguments{
\item{local_file}{Where the file is located on your computer.}

\item{path}{The destination on Cloudstor.}

\item{file_name}{Optional. What you want to call the file on Cloudstor?
If it is not provided, it is the same as the file name of the local file}

\item{user}{Optional. Your Cloudstor username.}

\item{password}{Optional. Your Cloudstor password.}
}
\value{
Nothing is returned. A success or error message is printed.
}
\description{
Save a file to Cloudstor. If the file already exists, it is
replaced.
}
