% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{cloud_put}
\alias{cloud_put}
\title{cloud_put}
\usage{
cloud_put(
  file_name,
  local_file,
  path = "",
  user = cloud_auth_user(),
  password = cloud_auth_pwd()
)
}
\arguments{
\item{file_name}{What you want to call the file on cloudstor}

\item{local_file}{Where the file is located}

\item{path}{path to file or folder}

\item{user}{Your cloudstor username}

\item{password}{Your cloudstor password}
}
\value{
nothing
}
\description{
Save a file to cloudstor.
}
