% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstoR-package.R
\docType{package}
\name{cloudstoR-package}
\alias{cloudstoR}
\alias{cloudstoR-package}
\title{cloudstoR: Simplifies Access to Cloudstor API}
\description{
Access Cloudstor via their WebDAV API. This package can read, write, and navigate Cloudstor from R.
}
\author{
\strong{Maintainer}: Taren Sanders \email{taren.sanders@acu.edu.au} (\href{https://orcid.org/0000-0002-4504-6008}{ORCID})

Authors:
\itemize{
  \item Philip Parker \email{philip.parker@acu.edu.au} (\href{https://orcid.org/0000-0002-4604-8566}{ORCID})
  \item Timothy Churches
}

}
\keyword{internal}
