% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudstorapi-access.R
\name{get_handle}
\alias{get_handle}
\title{get_handle}
\usage{
get_handle(user, password, reset = FALSE)
}
\arguments{
\item{user}{cloudstor username}

\item{password}{cloudstor password}

\item{reset}{remove the existing authentication and handle}
}
\value{
curl handle object
}
\description{
Return a handle for CURL to use. Not a user facing function
}
